This is where hell begins. Rename 'area' to 'areas' so that /help area (command) does not conflict with /help areas (submodule)
(cherry picked from commit 7804b3b0279cd09306fc68f344933493447841c1)
