+++ server/commands/__init__.py
@@ -95,7 +95,7 @@ def mod_only(area_owners=False):
 # There must be an __all__ in each module in order for reloading
 # to work properly.
 from .admin import *
-from .area import *
+from .areas import *
 from .casing import *
 from .character import *
 from .fun import *
