+++ server/commands/areas.py
